Add support for repatching with additional perms/features + lots of UI improvements

+++ mbf-agent/src/patching.rs
@@ -17,9 +17,14 @@ const LIB_UNITY_PATH: &str = "lib/arm64-v8a/libunity.so";
 const DIFF_DOWNLOAD_ATTEMPTS: u32 = 3;
 
 // Mods the currently installed version of the given app and reinstalls it, without doing any downgrading.
-pub fn mod_current_apk(temp_path: &Path, app_info: &AppInfo, manifest_mod: ManifestMod) -> Result<()> {
-    info!("Downloading unstripped libunity.so (this could take a minute)");
-    let libunity_path = save_libunity(temp_path, &app_info.version).context("Failed to save libunity.so")?;
+// If `manifest_only` is true, patching will only attempt to update permissions/features 
+pub fn mod_current_apk(temp_path: &Path, app_info: &AppInfo, manifest_mod: ManifestMod, manifest_only: bool) -> Result<()> {
+    let libunity_path = if manifest_only {
+        None
+    }   else    {
+        info!("Downloading unstripped libunity.so (this could take a minute)");
+        save_libunity(temp_path, &app_info.version).context("Failed to save libunity.so")?
+    };
 
     kill_app()?;
 
@@ -32,7 +37,7 @@ pub fn mod_current_apk(temp_path: &Path, app_info: &AppInfo, manifest_mod: Manif
     std::fs::create_dir(&obb_backup)?;
     let obb_backups = save_obbs(Path::new(APP_OBB_PATH), &obb_backup)?;
 
-    patch_and_reinstall(libunity_path, &temp_apk_path, temp_path, obb_backups, manifest_mod)?;
+    patch_and_reinstall(libunity_path, &temp_apk_path, temp_path, obb_backups, manifest_mod, manifest_only)?;
     Ok(())
 }
 
@@ -76,7 +81,7 @@ pub fn downgrade_and_mod_apk(temp_path: &Path,
         obb_backup_paths.push(obb_backup_path);
     }
 
-    patch_and_reinstall(libunity_path, &temp_apk_path, temp_path, obb_backup_paths, manifest_mod)?;
+    patch_and_reinstall(libunity_path, &temp_apk_path, temp_path, obb_backup_paths, manifest_mod, false)?;
     Ok(())
 }
 
@@ -92,9 +97,10 @@ fn patch_and_reinstall(libunity_path: Option<PathBuf>,
     temp_apk_path: &Path,
     temp_path: &Path,
     obb_paths: Vec<PathBuf>,
-    manifest_mod: ManifestMod) -> Result<()> {
+    manifest_mod: ManifestMod,
+    manifest_only: bool) -> Result<()> {
     info!("Patching APK at {:?}", temp_path);
-    patch_apk_in_place(&temp_apk_path, libunity_path, manifest_mod)?;
+    patch_apk_in_place(&temp_apk_path, libunity_path, manifest_mod, manifest_only)?;
 
     if Path::new(PLAYER_DATA_PATH).exists() {
         info!("Backing up player data");
@@ -334,7 +340,7 @@ pub fn install_modloader() -> Result<()> {
     Ok(())
 }
 
-fn patch_apk_in_place(path: impl AsRef<Path>, libunity_path: Option<PathBuf>, manifest_mod: ManifestMod) -> Result<()> {
+fn patch_apk_in_place(path: impl AsRef<Path>, libunity_path: Option<PathBuf>, manifest_mod: ManifestMod, manifest_only: bool) -> Result<()> {
     let file = OpenOptions::new()
         .read(true)
         .write(true)
@@ -348,23 +354,25 @@ fn patch_apk_in_place(path: impl AsRef<Path>, libunity_path: Option<PathBuf>, ma
 
     let (priv_key, cert) = signing::load_cert_and_priv_key(DEBUG_CERT_PEM);
 
-    info!("Adding libmainloader");
-    zip.delete_file(LIB_MAIN_PATH);
-    zip.write_file(LIB_MAIN_PATH, &mut Cursor::new(LIB_MAIN), FileCompression::Deflate)?;
-    add_modded_tag(&mut zip, ModTag {
-        patcher_name: "ModsBeforeFriday".to_string(),
-        patcher_version: Some("0.1.0".to_string()), // TODO: Get this from the frontend maybe?
-        modloader_name: "Scotland2".to_string(), // TODO: This should really be Libmainloader because SL2 isn't inside the APK
-        modloader_version: None // Temporary, but this field is universally considered to be option so this should be OK.
-    })?;
+    if !manifest_only {
+        info!("Adding libmainloader");
+        zip.delete_file(LIB_MAIN_PATH);
+        zip.write_file(LIB_MAIN_PATH, &mut Cursor::new(LIB_MAIN), FileCompression::Deflate)?;
+        add_modded_tag(&mut zip, ModTag {
+            patcher_name: "ModsBeforeFriday".to_string(),
+            patcher_version: Some("0.1.0".to_string()), // TODO: Get this from the frontend maybe?
+            modloader_name: "Scotland2".to_string(), // TODO: This should really be Libmainloader because SL2 isn't inside the APK
+            modloader_version: None // Temporary, but this field is universally considered to be option so this should be OK.
+        })?;
 
-    info!("Adding unstripped libunity.so");
-    match libunity_path {
-        Some(unity_path) => {
-            let mut unity_stream = File::open(unity_path)?;
-            zip.write_file(LIB_UNITY_PATH, &mut unity_stream, FileCompression::Deflate)?;
-        },
-        None => warn!("No unstripped unity added to the APK! This might cause issues later")
+        info!("Adding unstripped libunity.so");
+        match libunity_path {
+            Some(unity_path) => {
+                let mut unity_stream = File::open(unity_path)?;
+                zip.write_file(LIB_UNITY_PATH, &mut unity_stream, FileCompression::Deflate)?;
+            },
+            None => warn!("No unstripped unity added to the APK! This might cause issues later")
+        }
     }
 
     info!("Signing");
